// web/src/app/(client)/(client-components)/type.ts
export interface OccupancyObject {
  rooms: number;
  adults: number;
  children: number;
  infants: number;
}

export type StaySearchFormFields = "location" | "guests" | "dates";

export interface PropertyType {
  name: string;
  description: string;
  checked: boolean;
}

export interface ClassOfProperties extends PropertyType {}

export type DateRage = [Date | null, Date | null];